+++ native/src/main/java/io/ballerina/stdlib/file/service/FSListener.java
@@ -31,8 +31,6 @@ import io.ballerina.runtime.api.values.BObject;
 import io.ballerina.runtime.api.values.BString;
 import io.ballerina.stdlib.file.utils.FileConstants;
 import io.ballerina.stdlib.file.utils.ModuleUtils;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 import org.wso2.transport.localfilesystem.server.connector.contract.LocalFileSystemEvent;
 import org.wso2.transport.localfilesystem.server.connector.contract.LocalFileSystemListener;
 
@@ -46,7 +44,6 @@ import static io.ballerina.stdlib.file.service.DirectoryListenerConstants.FILE_S
  */
 public class FSListener implements LocalFileSystemListener {
 
-    private static final Logger log = LoggerFactory.getLogger(FSListener.class);
     private Runtime runtime;
     private Map<BObject, Map<String, MethodType>> serviceRegistry = new HashMap<>();
 
@@ -67,9 +64,8 @@ public class FSListener implements LocalFileSystemListener {
                     boolean isConcurrentSafe = type.isIsolated() && type.isIsolated(functionName);
                     Object result = runtime.callMethod(service, functionName,
                             new StrandMetadata(isConcurrentSafe, null), balFileEvent);
-                    if (result instanceof BError error) {
-                        log.error("error returned from the '{}' remote function in file listener service: {}",
-                                functionName, error.getMessage());
+                    if (result instanceof BError bError) {
+                        bError.printStackTrace();
                     }
                 }
             }